chore(request): move HttpRequest to http namespace
