+++ src/www/send.cpp
@@ -10,15 +10,13 @@
 #include <sstream>
 #include <vector>
 
+#include "../../include/http/HttpRequest.hpp"
 #include "../../include/www/Client.hpp"
-#include "../../include/www/HttpRequest.hpp"
 #include "../../include/www/ServerSocket.hpp"
 
 #define DEFAULT_CHUNK_SIZE 64
 const char* SEQUENCE = "\r\n\r\n";
 
-typedef std::pair<std::string, std::string> t_field;
-typedef std::map<t_field::first_type, t_field::second_type> t_fields;
 typedef std::vector<char> t_bytes;
 
 // if doesnt contain \r\n\r\n, continue on (return false)
@@ -74,8 +72,8 @@ t_bytes::iterator endOfToken(t_bytes::iterator begin, t_bytes::iterator end) {
 // Read each line until we find an empty line (which indicates the end of
 // headers)
 
-t_fields parseHeaders(std::string& receivedBytes) {
-  t_fields newHeaders;
+http::t_fields parseHeaders(std::string& receivedBytes) {
+  http::t_fields newHeaders;
   size_t seqEndIndex = receivedBytes.find(SEQUENCE, 0UL, sizeof(SEQUENCE));
   if (seqEndIndex == std::string::npos) {
     return newHeaders;  // don't yet have end of transmission
@@ -127,8 +125,8 @@ int doKoolShit(const int port) {
 
   char buffer[DEFAULT_CHUNK_SIZE];
   std::string received;
-  www::HttpRequest request;
-  t_fields currentHeaders = request.getHeaderFields();
+  http::HttpRequest request;
+  http::t_fields currentHeaders = request.getHeaderFields();
 
   while (true) {
     int bytes_read = 0;
@@ -163,7 +161,7 @@ int doKoolShit(const int port) {
     // blocks forever if total length % 4 == 0
   }
   std::cout << "Headers received:" << std::endl;
-  for (t_fields::iterator it = currentHeaders.begin(); it != currentHeaders.end(); ++it) {
+  for (http::t_fields::iterator it = currentHeaders.begin(); it != currentHeaders.end(); ++it) {
     std::cout << "- " << it->first << ": " << it->second << std::endl;
   }
   std::cout << std::endl;
